+++ CHANGELOG.md
@@ -14,6 +14,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - pconf - fixed allocation of array E for LAPACK subroutines
 - removed IPmr parameter from params.f90 and refactored IPmr-associated arrays
 - pconf - refactored KXIJ and KWeights into optional inputs
+- removed IP1 parameter from params.f90 and refactored IP1-associated arrays from main programs
 
 ## [1.0.7] - 2024-11-17
 - refactored IP1 parameter out of pdtm.f90 and removed IP1 array size warnings